@@ -2,7 +2,7 @@ name: ecosecha
 description: "A new Flutter project."
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.3.0-120.0.dev <4.0.0'
+  sdk: ">=3.3.0-120.0.dev <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -32,17 +32,22 @@ dependencies:
   cloud_firestore: ^4.13.1
   cupertino_icons: ^1.0.2
   dropdown_button2: ^2.3.9
+  email_validator: ^2.1.17
   firebase_auth: ^4.14.0
   firebase_core: ^2.22.0
   firebase_storage: ^11.5.1
   flutter:
     sdk: flutter
+  flutter_paypal_payment: ^1.0.0
+  font_awesome_flutter: ^10.7.0
   food_icons: ^0.0.6
+  hexcolor: ^3.0.1
   image_picker: ^1.0.4
   intl: ^0.18.1
+  provider: ^6.1.2
   raised_buttons: ^0.0.2
+  string_validator: ^1.0.2
   uuid: ^4.2.1
-  flutter_paypal_payment: ^1.0.0
 
 dev_dependencies:
   flutter_lints: ^3.0.0
@@ -53,7 +58,6 @@ dev_dependencies:
 # following page: https://dart.dev/tools/pub/pubspec
 # The following section is specific to Flutter packages.
 flutter:
-
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
@@ -62,6 +66,14 @@ flutter:
   assets:
     - assets/images/ecosecha_logo.png
     - assets/images/fondo.png
+    - assets/images/designf.jpg
+    - assets/images/designs.jpg
+    - assets/images/designt.jpg
+    - assets/images/OIG1.jpeg
+    - assets/images/OIG2.jpeg
+    - assets/images/OIG3.jpeg
+    - assets/images/OIG4.jpeg
+    - assets/images/OIG5.jpeg
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
   # For details regarding adding assets from package dependencies, see
